+++ redux-kotlin/src/commonMain/kotlin/org/reduxkotlin/Compose.kt
@@ -6,5 +6,8 @@ package org.reduxkotlin
 public fun <T> compose(vararg functions: (T) -> T): (T) -> T =
   { x -> functions.foldRight(x) { f, composed -> f(composed) } }
 
+/**
+ * Composes a list of single argument functions from right to left.
+ */
 public fun <T> compose(functions: List<(T) -> T>): (T) -> T =
   { x -> functions.foldRight(x) { f, composed -> f(composed) } }